requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
cloudinary
python-dotenv
paddleocr
Pillow 